# docs/assets/openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/net-worth/timeline/account/{id}": {
      "get": {
        "description": "Retrieves the net worth overtime for the specific given account",
        "operationId": "NetWorthController_getNetWorthTimelineAccount",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Net worth over time successfully generated.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HistoricalDataPoint"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get net worth over time (timeline) of a specific account.",
        "tags": ["Net Worth"]
      }
    },
...
    "/account/{id}": {
      "get": {
        "description": "Retrieves an account by the given ID.",
        "operationId": "AccountController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account found successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required."
          },
          "404": {
            "description": "Account with the specified ID not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get account by ID.",
        "tags": ["Account"]
      },
      "delete": {
        "description": "Deletes an account by the given ID.",
        "operationId": "AccountController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account deleted successfully."
          },
          "401": {
            "description": "Authentication is required."
          },
          "404": {
            "description": "Account with the specified ID not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete account by ID.",
        "tags": ["Account"]
      },
      "patch": {
        "description": "Edits an account by the given ID.",
        "operationId": "AccountController_edit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountEditRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required."
          },
          "404": {
            "description": "Account with the specified ID not found or does not belong to the user."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Edit account.",
        "tags": ["Account"]
      }
    },
    "/account": {
      "get": {
        "description": "Retrieves all accounts for the authenticated user.",
        "operationId": "AccountController_getAccounts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Accounts found successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Account"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get accounts.",
        "tags": ["Account"]
      }
    },
    "/account/{id}/merge": {
      "post": {
        "description": "Merges a source account into the target account by Id, updating all related historical data and deleting the source. Intended purely to migrate from one account as your base to another, in the event the provider changes the structure. You should consider the Target Id (in the query) will be the remaining account. The source account will be provided by the body.",
        "operationId": "AccountController_mergeAccounts",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountMergeDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Accounts merged successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required."
          },
          "404": {
            "description": "One or both accounts not found or do not belong to the user."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Merge two accounts.",
        "tags": ["Account"]
      }
    },
...
  },
  "info": {
    "title": "Sprout API",
    "description": "Welcome to the Sprout API documentation. This document provides a comprehensive guide to all available endpoints. Use the sections below to explore different parts of the API.\n\n  **Note:** If you wish to test the functionality, you must use a login endpoint to properly attach cookies to your browser.",
    "version": "v0.5.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "ProviderTypeEnum": {
        "type": "string",
        "enum": ["simple-fin", "zillow", "plaid", "snapTrade"]
      },
...
      "AccountTypeEnum": {
        "type": "string",
        "enum": ["other", "depository", "credit", "asset", "loan", "investment", "crypto"],
        "description": "The type of this account to better separate it from the others."
      },
      "AccountSubTypeEnum": {
        "type": "string",
        "enum": [
          "other",
          "Savings",
          "Checking",
          "HYSA",
          "401K",
          "Brokerage",
          "IRA",
          "HSA",
          "Student",
          "Mortgage",
          "Personal",
          "Auto",
          "Travel",
          "Cash Back",
          "Wallet",
          "Staking",
          "House"
        ],
        "description": "The subtype of this account. For example, a depository could be a checking account, savings account, or HYSA."
      },
      "InstitutionIconType": {
        "type": "string",
        "enum": ["icon", "symbol"],
        "description": "The asset variant variant to request from Brandfetch (icon or symbol)"
      },
      "Institution": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "iconType": {
            "description": "The asset variant variant to request from Brandfetch (icon or symbol)",
            "default": "icon",
            "allOf": [
              {
                "$ref": "#/components/schemas/InstitutionIconType"
              }
            ]
          },
          "url": {
            "type": "string",
            "description": "The URL for where this institution is"
          },
          "name": {
            "type": "string"
          },
          "hasError": {
            "type": "boolean",
            "description": "If this institution has connection errors and needs fixed"
          }
        },
        "required": ["id", "iconType", "url", "name", "hasError"]
      },
      "Account": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "provider": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ProviderTypeEnum"
              }
            ]
          },
          "type": {
            "description": "The type of this account to better separate it from the others.",
            "allOf": [
              {
                "$ref": "#/components/schemas/AccountTypeEnum"
              }
            ]
          },
          "subType": {
            "description": "The subtype of this account. For example, a depository could be a checking account, savings account, or HYSA.",
            "allOf": [
              {
                "$ref": "#/components/schemas/AccountSubTypeEnum"
              }
            ]
          },
          "interestRate": {
            "type": "number",
            "description": "An interest rate if this is a loan type account.",
            "nullable": true
          },
          "balance": {
            "type": "number",
            "description": "The numeric value converted to the user's preferred currency format. This overrides the original balance property.",
            "nullable": false
          },
          "name": {
            "type": "string"
          },
          "institution": {
            "description": "The institution associated to this account",
            "allOf": [
              {
                "$ref": "#/components/schemas/Institution"
              }
            ]
          },
          "extra": {
            "type": "object",
            "description": "Any extra data that we want to store as JSON"
          }
        },
        "required": ["id", "provider", "type", "balance", "name", "institution"]
      },
...
      "HistoricalDataPoint": {
        "type": "object",
        "properties": {
          "date": {
            "format": "date-time",
            "type": "string",
            "description": "The date of the record"
          },
          "value": {
            "type": "number",
            "description": "The numerical currency value"
          }
        },
        "required": ["date", "value"]
      },
...
      "AccountEditRequest": {
        "type": "object",
        "properties": {
          "subType": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AccountSubTypeEnum"
              }
            ]
          },
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AccountTypeEnum"
              }
            ]
          },
          "name": {
            "type": "string"
          },
          "interestRate": {
            "type": "number"
          }
        }
      },
      "AccountMergeDTO": {
        "type": "object",
        "properties": {
          "sourceId": {
            "type": "string",
            "description": "The ID of the source account to merge and delete."
          }
        },
        "required": ["sourceId"]
      },
...
    }
  }
}